+++ mixins/policy-TerraformUpdateAccess.tf
@@ -1,11 +1,5 @@
-variable "tfstate_environment_name" {
-  type        = string
-  description = "The name of the environment where `tfstate-backend` is provisioned. If not set, the TerraformUpdateAccess permission set will not be created."
-  default     = null
-}
-
 locals {
-  tf_update_access_enabled = var.tfstate_environment_name != null && module.this.enabled
+  tf_update_access_enabled = var.tfstate_environment_name != null && local.enabled
 }
 
 module "tfstate" {
@@ -29,14 +23,14 @@ data "aws_iam_policy_document" "terraform_update_access" {
     sid     = "TerraformStateBackendS3Bucket"
     effect  = "Allow"
     actions = ["s3:ListBucket", "s3:GetObject", "s3:PutObject"]
-    resources = module.this.enabled ? [
+    resources = local.enabled ? [
       module.tfstate.outputs.tfstate_backend_s3_bucket_arn,
       "${module.tfstate.outputs.tfstate_backend_s3_bucket_arn}/*"
     ] : []
   }
 
   dynamic "statement" {
-    for_each = (module.this.enabled && module.tfstate.outputs.tfstate_backend_dynamodb_table_arn != "") ? [1] : []
+    for_each = (local.enabled && module.tfstate.outputs.tfstate_backend_dynamodb_table_arn != "") ? [1] : []
 
     content {
       sid       = "TerraformStateBackendDynamoDbTable"