+++ src/policy-BillingReadOnlyAccess.tf
@@ -1,5 +1,5 @@
 locals {
-  billing_read_only_access_permission_set = [{
+  billing_read_only_access_permission_set = local.enabled ? [{
     name             = "BillingReadOnlyAccess",
     description      = "Allow users to view bills in the billing console",
     relay_state      = "",
@@ -11,5 +11,5 @@ locals {
       "arn:${local.aws_partition}:iam::aws:policy/AWSSupportAccess",
     ]
     customer_managed_policy_attachments = []
-  }]
+  }] : []
 }